+++ src/pages/DashboardAdmin/components/Personnel/AddAdminPersonnel.tsx
@@ -1,25 +1,20 @@
-import { FixedBox, LoginTitle, BackIcon } from "../styles/AddPersonnel.style";
-import { registerDoctor } from "../../../api/users/index";
+import {
+  FixedBox,
+  LoginTitle,
+  BackIcon,
+} from "../../styles/AddPersonnel.style";
+import { registerAdmin } from "../../../../api/users/index";
 import { useEffect, useState } from "react";
 import { useForm } from "react-hook-form";
-import DocumentType from "./constants/document-type.json";
+import DocumentType from "../constants/document-type.json";
+import { INewAdmin } from "../../../../models/IAdmin";
 import { Button } from "react-bootstrap";
 import { useHistory } from "react-router";
-import { ConfirmationModal } from "../../../components/ConfirmationModal";
-import ArrowLeft from "../../../resources/ArrowLeft.svg";
-import { getSpecialtyByHospital } from "../../../api/utilities";
-import { INewDoctor } from "../../../models/IDoctor";
+import { ConfirmationModal } from "../../../../components/ConfirmationModal";
+import ArrowLeft from "../../../../resources/ArrowLeft.svg";
 
-type SelectSpecialty = {
-  label: string;
-  value: number;
-};
-
-const AddDoctorPersonnel = () => {
+const AddAdminPersonnel = () => {
   const [isShowModal, setIsShowModal] = useState(false);
-  const [specialties, setSpecialties] = useState<SelectSpecialty[]>([
-    { label: "", value: 0 },
-  ]);
   const history = useHistory();
   const { register, handleSubmit, reset } = useForm({
     defaultValues: {
@@ -29,10 +24,8 @@ const AddDoctorPersonnel = () => {
       email: "",
       document_number: "",
       date_of_birth: new Date(),
-      doctor: {
+      admin: {
         hospital_id: 0,
-        schedule: "",
-        specialties: [0],
       },
     },
   });
@@ -49,45 +42,28 @@ const AddDoctorPersonnel = () => {
     history.goBack();
   };
 
-  const getSpecialties = async () => {
-    const hospitalSpecialties = await getSpecialtyByHospital(
-      Number(localStorage.getItem("hospitalId"))
-    );
-    const allSpecialties = hospitalSpecialties.data.map((item) => {
-      return {
-        label: item.name,
-        value: item.id,
-      };
-    });
-    setSpecialties(allSpecialties);
-  };
-
   useEffect(() => {
     const fetchData = async () => {
-      // This would be a GET call to an endpoint
       reset({
         document_type: "Tipo de Documento",
       });
     };
 
     fetchData();
-    getSpecialties();
   }, [reset]);
 
   const onSubmit = async (data: any) => {
     try {
-      const doctorData: INewDoctor = {
-        user_role: "doctor",
+      const adminData: INewAdmin = {
+        user_role: "admin",
         ...data,
-        doctor: {
+        admin: {
           hospital_id: Number(localStorage.getItem("hospitalId")),
-          schedule: data.doctor.schedule,
-          specialties: data.doctor.specialties,
         },
       };
 
-      if (doctorData) {
-        registerDoctor(doctorData)
+      if (adminData) {
+        registerAdmin(adminData)
           .catch((err) => {
             console.log(err);
           })
@@ -112,14 +88,14 @@ const AddDoctorPersonnel = () => {
       <div className="d-flex justify-content-center align-items-center h-100">
         <FixedBox>
           <form
-            id="add-doctor-form"
+            id="add-admin-form"
             className="row d-flex h-100"
             onSubmit={handleSubmit(onSubmit)}
           >
             <div className="col-7 h-100 w-100">
               <div className="row d-flex h-100 align-items-center m-2 pt-5">
                 <form className="d-flex flex-column w-100 pe-3">
-                  <LoginTitle>Registrar Doctor</LoginTitle>
+                  <LoginTitle>Registrar Admin</LoginTitle>
                   <div className="form-group d-flex justify-content-start mb-2 me-1 w-100">
                     <input
                       type="text"
@@ -175,26 +151,8 @@ const AddDoctorPersonnel = () => {
                       {...register("date_of_birth", { required: true })}
                     />
                   </div>
-                  <div className="form-group d-flex justify-content-start mb-2">
-                    <input
-                      type="text"
-                      className="form-control"
-                      placeholder="Horario"
-                      {...register("doctor.schedule", { required: true })}
-                    />
-                  </div>
-                  <select
-                    className="form-select"
-                    multiple
-                    {...register("doctor.specialties", { required: true })}
-                  >
-                    {specialties.map((option) => {
-                      return (
-                        <option value={option.value}>{option.label}</option>
-                      );
-                    })}
-                  </select>
-                  <div className="form-group d-flex justify-content-end mb-3 mt-4">
+
+                  <div className="form-group d-flex justify-content-end mb-3">
                     <Button variant="primary" onClick={showModal}>
                       Registrar
                     </Button>
@@ -210,7 +168,7 @@ const AddDoctorPersonnel = () => {
                 button1Text="Cancelar"
                 button2Text="Confirmar"
                 handleShow={updateModal}
-                formId={"add-doctor-form"}
+                formId={"add-admin-form"}
               />
             )}
           </form>
@@ -220,4 +178,4 @@ const AddDoctorPersonnel = () => {
   );
 };
 
-export default AddDoctorPersonnel;
+export default AddAdminPersonnel;
