+++ src/pages/DashboardAdmin/components/Personnel/AddDoctorPersonnel.tsx
@@ -1,16 +1,29 @@
-import { FixedBox, LoginTitle, BackIcon } from "../styles/AddPersonnel.style";
-import { registerAdmin } from "../../../api/users/index";
+import {
+  FixedBox,
+  LoginTitle,
+  BackIcon,
+} from "../../styles/AddPersonnel.style";
+import { registerDoctor } from "../../../../api/users/index";
 import { useEffect, useState } from "react";
 import { useForm } from "react-hook-form";
-import DocumentType from "./constants/document-type.json";
-import { INewAdmin } from "../../../models/IAdmin";
+import DocumentType from "../constants/document-type.json";
 import { Button } from "react-bootstrap";
 import { useHistory } from "react-router";
-import { ConfirmationModal } from "../../../components/ConfirmationModal";
-import ArrowLeft from "../../../resources/ArrowLeft.svg";
+import { ConfirmationModal } from "../../../../components/ConfirmationModal";
+import ArrowLeft from "../../../../resources/ArrowLeft.svg";
+import { getSpecialtyByHospital } from "../../../../api/utilities";
+import { INewDoctor } from "../../../../models/IDoctor";
 
-const AddAdminPersonnel = () => {
+type SelectSpecialty = {
+  label: string;
+  value: number;
+};
+
+const AddDoctorPersonnel = () => {
   const [isShowModal, setIsShowModal] = useState(false);
+  const [specialties, setSpecialties] = useState<SelectSpecialty[]>([
+    { label: "", value: 0 },
+  ]);
   const history = useHistory();
   const { register, handleSubmit, reset } = useForm({
     defaultValues: {
@@ -20,8 +33,10 @@ const AddAdminPersonnel = () => {
       email: "",
       document_number: "",
       date_of_birth: new Date(),
-      admin: {
+      doctor: {
         hospital_id: 0,
+        schedule: "",
+        specialties: [0],
       },
     },
   });
@@ -38,28 +53,45 @@ const AddAdminPersonnel = () => {
     history.goBack();
   };
 
+  const getSpecialties = async () => {
+    const hospitalSpecialties = await getSpecialtyByHospital(
+      Number(localStorage.getItem("hospitalId"))
+    );
+    const allSpecialties = hospitalSpecialties.data.map((item) => {
+      return {
+        label: item.name,
+        value: item.id,
+      };
+    });
+    setSpecialties(allSpecialties);
+  };
+
   useEffect(() => {
     const fetchData = async () => {
+      // This would be a GET call to an endpoint
       reset({
         document_type: "Tipo de Documento",
       });
     };
 
     fetchData();
+    getSpecialties();
   }, [reset]);
 
   const onSubmit = async (data: any) => {
     try {
-      const adminData: INewAdmin = {
-        user_role: "admin",
+      const doctorData: INewDoctor = {
+        user_role: "doctor",
         ...data,
-        admin: {
+        doctor: {
           hospital_id: Number(localStorage.getItem("hospitalId")),
+          schedule: data.doctor.schedule,
+          specialties: data.doctor.specialties,
         },
       };
 
-      if (adminData) {
-        registerAdmin(adminData)
+      if (doctorData) {
+        registerDoctor(doctorData)
           .catch((err) => {
             console.log(err);
           })
@@ -84,14 +116,14 @@ const AddAdminPersonnel = () => {
       <div className="d-flex justify-content-center align-items-center h-100">
         <FixedBox>
           <form
-            id="add-admin-form"
+            id="add-doctor-form"
             className="row d-flex h-100"
             onSubmit={handleSubmit(onSubmit)}
           >
             <div className="col-7 h-100 w-100">
               <div className="row d-flex h-100 align-items-center m-2 pt-5">
                 <form className="d-flex flex-column w-100 pe-3">
-                  <LoginTitle>Registrar Admin</LoginTitle>
+                  <LoginTitle>Registrar Doctor</LoginTitle>
                   <div className="form-group d-flex justify-content-start mb-2 me-1 w-100">
                     <input
                       type="text"
@@ -147,8 +179,26 @@ const AddAdminPersonnel = () => {
                       {...register("date_of_birth", { required: true })}
                     />
                   </div>
-
-                  <div className="form-group d-flex justify-content-end mb-3">
+                  <div className="form-group d-flex justify-content-start mb-2">
+                    <input
+                      type="text"
+                      className="form-control"
+                      placeholder="Horario"
+                      {...register("doctor.schedule", { required: true })}
+                    />
+                  </div>
+                  <select
+                    className="form-select"
+                    multiple
+                    {...register("doctor.specialties", { required: true })}
+                  >
+                    {specialties.map((option) => {
+                      return (
+                        <option value={option.value}>{option.label}</option>
+                      );
+                    })}
+                  </select>
+                  <div className="form-group d-flex justify-content-end mb-3 mt-4">
                     <Button variant="primary" onClick={showModal}>
                       Registrar
                     </Button>
@@ -164,7 +214,7 @@ const AddAdminPersonnel = () => {
                 button1Text="Cancelar"
                 button2Text="Confirmar"
                 handleShow={updateModal}
-                formId={"add-admin-form"}
+                formId={"add-doctor-form"}
               />
             )}
           </form>
@@ -174,4 +224,4 @@ const AddAdminPersonnel = () => {
   );
 };
 
-export default AddAdminPersonnel;
+export default AddDoctorPersonnel;
